move CMS requests to top level app
