--- src/app/home/home.component.html
@@ -1,1 +1,0 @@
-<metrics-downloads *ngIf="filter"></metrics-downloads>
