+++ src/app/home/home.component.ts
@@ -1,120 +1,12 @@
-import { Component, OnInit } from '@angular/core';
-import { Store } from '@ngrx/store';
-import { Observable } from 'rxjs/Observable';
-import { CmsService } from '../core';
-import { Env } from '../core/core.env';
-import { HalDoc } from 'ngx-prx-styleguide';
-import { EpisodeModel, PodcastModel, FilterModel } from '../ngrx/model';
-import { cmsPodcastFeed, cmsEpisodeGuid } from '../ngrx/actions/cms.action.creator';
-import { castleFilter } from '../ngrx/actions/castle.action.creator';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'metrics-home',
-  templateUrl: './home.component.html',
-  styleUrls: ['./home.component.css']
+  template: `
+    <metrics-downloads></metrics-downloads>
+  `
 })
-export class HomeComponent implements OnInit {
-  podcastStore: Observable<PodcastModel[]>;
-  podcasts: PodcastModel[];
-  filterStore: Observable<FilterModel[]>;
-  filter: FilterModel;
-
-  constructor(private cms: CmsService, private store: Store<any>) {
-    this.podcastStore = store.select('podcast');
-    this.filterStore = store.select('filter');
-  }
-
-  ngOnInit() {
-    this.cms.auth.subscribe(auth => {
-      auth.followItems('prx:series', {filters: 'v4'}).subscribe((series: HalDoc[]) => {
-        series.map(doc => {
-          return {
-            doc,
-            seriesId: doc['id'],
-            title: doc['title']
-          };
-        }).forEach(this.getSeriesPodcastDistribution.bind(this));
-      });
-    });
-
-    this.podcastStore.subscribe((state: PodcastModel[]) => {
-      this.podcasts = state;
-
-      if (this.podcasts.length > 0) {
-        let selectedPodcast;
-        if (Env.CASTLE_TEST_PODCAST) {
-          selectedPodcast = this.podcasts.find((p: PodcastModel) => {
-            return p.feederId === Env.CASTLE_TEST_PODCAST.toString();
-          });
-        } else {
-          selectedPodcast = this.podcasts;
-        }
-        if (selectedPodcast &&
-          (!this.filter || (this.filter && this.filter.podcast && selectedPodcast.seriesId !== this.filter.podcast.seriesId))) {
-          this.store.dispatch(castleFilter({podcast: selectedPodcast}));
-        }
-      }
-    });
-
-    this.filterStore.subscribe((state: FilterModel) => {
-      if (state.podcast && (!this.filter || state.podcast.seriesId !== this.filter.podcast.seriesId)) {
-        this.getEpisodes(state.podcast);
-        this.filter = state;
-      }
-    });
-  }
-
-  getSeriesPodcastDistribution(podcast: PodcastModel) {
-    podcast.doc.followItems('prx:distributions').subscribe((distros: HalDoc[]) => {
-      const podcasts = distros.filter((doc => doc['kind'] === 'podcast'));
-      if (podcasts && podcasts.length > 0) {
-        podcast.feederUrl = podcasts[0]['url']; // TODO: am I supposed to get the feeder id from this url?
-        const urlParts = podcast.feederUrl.split('/');
-        if (urlParts.length > 1) {
-          podcast.feederId = urlParts[urlParts.length - 1];
-
-          this.store.dispatch(cmsPodcastFeed(podcast));
-        }
-      }
-    });
-  }
-
-  getEpisodes(podcast: PodcastModel) {
-    podcast.doc.followItems('prx:stories', {
-      per: podcast.doc.count('prx:stories'),
-      filters: 'v4',
-      sorts: 'released_at: desc, published_at: desc'
-    }).subscribe((episodes: HalDoc[]) => {
-      episodes.map(doc => {
-        return {
-          doc,
-          id: doc['id'],
-          seriesId: podcast.seriesId,
-          title: doc['title'],
-          publishedAt: doc['publishedAt'] ? new Date(doc['publishedAt']) : null
-        };
-      }).forEach((e) => {
-        // only include episodes with publish dates
-        if (e.publishedAt) {
-          this.getEpisodePodcastDistribution(podcast, e);
-        }
-      });
-    });
-  }
-
-  getEpisodePodcastDistribution(podcast: PodcastModel, episode: EpisodeModel) {
-    episode.doc.followItems('prx:distributions').subscribe((distros: HalDoc[]) => {
-      const podcasts = distros.filter((doc => doc['kind'] === 'episode'));
-      if (podcasts && podcasts.length > 0) {
-        episode.feederUrl = podcasts[0]['url'];
-
-        const urlParts = episode.feederUrl.split('/');
-        if (urlParts.length > 1) {
-          episode.guid = urlParts[urlParts.length - 1];
-
-          this.store.dispatch(cmsEpisodeGuid(podcast, episode));
-        }
-      }
-    });
-  }
+export class HomeComponent {
+  // eventually this will have Tabs+TabService (or whatever else we use because of the mobile requirement),
+  //  but right now it does nothing but a place to point the router and put downloads
 }
