+++ src/app/ngrx/reducers/episode-metrics.reducer.ts
@@ -24,7 +24,7 @@ export function EpisodeMetricsReducer(state: EpisodeMetricsModel[] = initialStat
         episode[metricsProperty] = unsparseDataset(action.payload.filter, action.payload.metrics);
         newState = [episode, ...state];
       }
-      console.log('EpisodeMetricsReducer', action.type, newState);
+      // console.log('EpisodeMetricsReducer', action.type, newState);
       return newState;
     default:
       return state;
